Do not use advisory locks if driver implements Lockable
If a driver implements the Lockable interface, then we
already use a manual table lock and hold it until all
the migrations are over. So there's no need to additionally
acquire advisory locks. They are just redundant.

Therefore, we maintain two modes of operation. Either a driver
implements Lockable, in which case we avoid advisory locks.
Or a driver does not do that, in which case we do use advisory locks
to do the locking for us.

It is up to the driver implementor to do as they wish.