Replace the SchemaName with the current schema
In postgres installations, sometimes the table
can be under a different schema. In that case,
we need to be able to access the information_schema.columns
table amongst others and filter by the table_schema.

This PR allows us to replace the {{.SchemaName}}
template in the SQL files with the actual
schema name.