MM-56321: Improve logging on lock acquisition failure (#32)
Sometimes, we observe that the line "lock.go:80: DB is locked, going to try acquire the lock if it is expired."
appears in a loop and it never proceeds. (Internal ref: ZD-42529)

There is a bug somewhere in morph due to which we are unable to release the lock automatically.
The only solution remains to manually remove the row from the db_lock table which is not ideal.

We need to find the bug, but without logging the actual error we would never know. So logging
the error is the first step.

https://mattermost.atlassian.net/browse/MM-56321